Pull repo.

+++ src/deb/repository.rs
@@ -153,7 +153,7 @@ pub struct ExtendedControlData {
     pub control: Package,
     hash: MultiHash,
     filename: PathBuf,
-    size: usize,
+    size: u64,
 }
 
 impl Display for ExtendedControlData {
